+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.3] - 2020-06-29
+
+### Added
+- Notification in Usermanagement Module
+
+### Changed
+- Multiuploader can't upload images
+- Changes Notification System
+
 ## [3.2] - 2020-06-27
 
 ### Changed